Fixed dealing undos, fixed scores, fixed cards left.
Dealing undo now actually deals in reverse when undoing, instead of that
weird bullshit where it dealt forwards both ways then got fixed on deck
flips.

The score of the game is now fully implemented and needs more testing.

The number of cards left in the main desk is now fully implemented and
needs more testing.